a lot of `usize` tweaks

+++ rts/motoko-rts/src/gc.rs
@@ -18,8 +18,8 @@ unsafe fn should_do_gc(max_live: crate::types::Bytes<u64>) -> bool {
     const HEAP_GROWTH_FACTOR: f64 = 1.5;
 
     let heap_limit = core::cmp::min(
-        (f64::from(LAST_HP) * HEAP_GROWTH_FACTOR) as u64,
-        (u64::from(LAST_HP) + max_live.0) / 2,
+        (f64::from(LAST_HP as u32) * HEAP_GROWTH_FACTOR) as u64,
+        (u64::from(LAST_HP as u32) + max_live.0) / 2,
     );
 
     u64::from(get_hp_unskewed() as u32) >= heap_limit
